enhance dashboard, new API, edit type data for gaji

+++ src/main/java/com/example/hrisapi/dto/response/KaryawanByNipResponse.java
@@ -35,7 +35,7 @@ public class KaryawanByNipResponse {
 
     private String email;
 
-    private Double gaji;
+    private Long gaji;
 
     private String gender;
